% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_log_table.R
\name{add_log_table}
\alias{add_log_table}
\title{Add a data.frame to a logging project}
\usage{
add_log_table(logging_project, log_id, log_table)
}
\arguments{
\item{logging_project}{A logging object from the `use_logging()` function}

\item{log_id}{The id value in the yaml file}

\item{log_table}{A data.frame to be logged}
}
\description{
Add a data.frame to a logging project
}
\examples{
start_logging(project = "my logging project",
                        log_file = "my_project_log.xlsx",
                        yaml_file = "my_log.yaml")
logging_project <- use_logging("my_log.yaml")
x <- data.frame(this = "foo", that = "bar")
logging_project <- add_log_table(logging_project, "1", x)
}
