% File src/R/man/levels.Rd
% $Id: levels.Rd 96 2011-09-20 08:39:31Z mariotomo $

\name{loglevels}
\alias{loglevels}
\title{the logging levels, names and values}
\description{
  
  this list associates names to values and vice versa.

  names and values are the same as in the python standard logging module.

}
\examples{
loglevels['INFO']
loglevels['DEBUG']
}