% File src/R/man/logging-internal.Rd
% $Id: logging-internal.Rd 97 2011-09-21 14:23:04Z mariotomo $

\name{logging-internal}
\alias{logging-internal}
\title{undocumented internal functions and variables}
\description{
  you hit this page because the function or variable
  has no real documentation page.
}
\alias{SVNVERSION}
\alias{SVNDATE}
\alias{updateOptions}
\alias{updateOptions.environment}
\alias{levellog}
\examples{
## no examples for undocumented functions
}