% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_logging.R
\name{save_logging}
\alias{save_logging}
\title{Save an xlsx file with logging tables}
\usage{
save_logging(logging_project)
}
\arguments{
\item{logging_project}{A logging object from the `use_logging()` function}
}
\description{
Save an xlsx file with logging tables
}
\examples{
start_logging(project = "my logging project",
                        log_file = "my_project_log.xlsx",
                        yaml_file = "my_log.yaml")
logging_project <- use_logging("my_log.yaml")
x <- data.frame(this = "foo", that = "bar")
logging_project <- add_log_table(logging_project, "1", x)
save_logging(logging_project)
}
