% File src/R/man/setLevel.Rd
% $Id: setLevel.Rd 11 2010-04-03 18:54:20Z mariotomo $

\name{setLevel}
\alias{setLevel}
\title{set logging.level for the named logger}
\description{

  alter an existing logger, setting its logging.level to a new value.

}
\usage{
setLevel(name, level)
}
\arguments{
  \item{name}{the name of the logger.}
  \item{level}{the new level for this logger.}
}
