% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logging.R
\name{start_logging}
\alias{start_logging}
\title{Create a log yaml file}
\usage{
start_logging(project, log_file, yaml_file)
}
\arguments{
\item{project}{The name of the project}

\item{log_file}{The full path of where the .xlsx Excel log file will be}

\item{yaml_file}{The full path of where the .yaml file will be}
}
\description{
Create a log yaml file
}
