% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_logging.R
\name{use_logging}
\alias{use_logging}
\title{Use a logging project}
\usage{
use_logging(yaml_file)
}
\arguments{
\item{yaml_file}{The full path of the .yaml file for the logging project}
}
\description{
Use a logging project
}
