% File src/R/man/levels.Rd
% $Id: levels.Rd 11 2010-04-03 18:54:20Z mariotomo $

\name{loglevels}
\alias{loglevels}
\title{the logging levels, names and values}
\description{
  
  this list associates names to values and vice versa.

  names and values are the same as in the python standard logging module.

}
