% File src/R/man/logging-internal.Rd
% $Id: logging-internal.Rd 18 2010-04-08 09:44:43Z mariotomo $

\name{logging-internal}
\title{undocumented internal functions and variables}
\description{
  you hit this page because the function or variable
  has no real documentation page.
}
\alias{SVNVERSION}
\alias{SVNDATE}
\alias{updateOptions}
\alias{updateOptions.environment}
\alias{levellog}
