% File src/R/man/logging-internal.Rd
% $Id: logging-internal.Rd 11 2010-04-03 18:54:20Z mariotomo $

\name{logging-internal}
\title{undocumented internal functions and variables}
\description{
  you hit this page because the function or variable
  has no real documentation page.
}
\alias{SVNVERSION}
\alias{SVNDATE}
