% File src/R/man/logging.addHandler.Rd

\name{logging.module}
\alias{logging.module}
\title{a logging module}
\description{
  
  What you find here behaves similarly to what you also find in Python.
  This includes hierarchic loggers, multiple handlers at each logger,
  the possibility to specify a formatter for each handler (one default
  formatter is given), same levels (names and numeric values) as
  Python's logging package, a simple logging.BasicConfig function to
  quickly put yourself in a usable situation...

  known to be missing:
  \itemize{
    \item{-}{setting a new formatter in an existing handler.}
    }
  
}

